Community components seperated
