+++ app/components/Community/CommunityHeader/CommunityHeader.jsx
@@ -14,8 +14,10 @@ import {
 import CommunityImage from "./CommunityImage";
 import { useEffect, useState } from "react";
 import axios from "axios";
+import CommunityNotFound from "./CommunityNotFound";
+import CommunityPageLoading from "../../Loading/Communities/CommunityPage/CommunityPageLoading";
 
-const Header = () => {
+const Header = ({ communityExists }) => {
   const toast = useToast();
 
   // States
@@ -103,17 +105,13 @@ const Header = () => {
         }
       });
 
-      console.log("Updated communities: ", updatedCommunities);
       const tempCommunities = localStorage.getItem("tempCommunities");
-      console.log("Temp:", tempCommunities);
       const resultArray = Object.assign(
         JSON.parse(tempCommunities),
         // @ts-ignore
         updatedCommunities
       );
-      console.log("To be saved to temp: ", resultArray);
       localStorage.setItem("tempCommunities", JSON.stringify(resultArray));
-      console.log(localStorage.getItem("tempCommunities"));
 
       // @ts-ignore
       setUserCommunityInfo(updatedCommunities);
@@ -131,101 +129,120 @@ const Header = () => {
   }, [user.authenticated]);
 
   return (
-    <Flex direction="row">
-      <Flex
-        _dark={{ bg: "customGray" }}
-        borderBottomRadius="0"
-        w="full"
-        h="160px"
-      >
-        <Flex flex="1" align="center" mb="8">
-          <Flex flex="1" mt="10" direction="row" justify="flex-start" ml="8">
-            <CommunityImage />
-            <Flex display={{ base: "none", sm: "block" }} direction="column">
-              <Text
-                fontSize="xl"
-                ml="2"
-                mb="1"
-                maxW={{ base: "200px", md: "400px", lg: "none" }}
-                noOfLines={2}
-                textOverflow="ellipsis"
+    <>
+      {communityExists === "yes" ? (
+        <Flex direction="row">
+          <Flex
+            _dark={{ bg: "customGray" }}
+            borderBottomRadius="0"
+            w="full"
+            h="160px"
+          >
+            <Flex flex="1" align="center" mb="8">
+              <Flex
+                flex="1"
+                mt="10"
+                direction="row"
+                justify="flex-start"
+                ml="8"
               >
-                {communityData.CommunityName}
-              </Text>
-              <Text
-                maxW="200px"
-                noOfLines={1}
-                fontSize="sm"
-                fontWeight="thin"
-                ml="3"
-              >
-                {communityData.CommunityName}
-              </Text>
-            </Flex>
-            <Flex flex="1" mr="10" justify="flex-end" align="center">
-              {user.authenticated && userCommunityInfo.isJoined ? (
-                <>
-                  <CommunitySettingsModal />
-                  <IconButton
-                    onClick={() => {
-                      setcommunitySettingsModal((prev) => ({
-                        ...prev,
-                        openCommunitySettingsModal: true,
-                      }));
-                    }}
-                    aria-label="Community Settings"
-                    mr="3"
-                    size="md"
-                    icon={<CustomCommunitySettingsIcon />}
-                  />
-                </>
-              ) : null}
+                <CommunityImage />
+                <Flex
+                  display={{ base: "none", sm: "block" }}
+                  direction="column"
+                >
+                  <Text
+                    fontSize="xl"
+                    ml="2"
+                    mb="1"
+                    maxW={{ base: "200px", md: "400px", lg: "none" }}
+                    noOfLines={2}
+                    textOverflow="ellipsis"
+                  >
+                    {communityData.CommunityName}
+                  </Text>
+                  <Text
+                    maxW="200px"
+                    noOfLines={1}
+                    fontSize="sm"
+                    fontWeight="thin"
+                    ml="3"
+                  >
+                    {communityData.CommunityName}
+                  </Text>
+                </Flex>
+                <Flex flex="1" mr="10" justify="flex-end" align="center">
+                  {user.authenticated && userCommunityInfo.isJoined ? (
+                    <>
+                      <CommunitySettingsModal />
+                      <IconButton
+                        onClick={() => {
+                          setcommunitySettingsModal((prev) => ({
+                            ...prev,
+                            openCommunitySettingsModal: true,
+                          }));
+                        }}
+                        aria-label="Community Settings"
+                        mr="3"
+                        size="md"
+                        icon={<CustomCommunitySettingsIcon />}
+                      />
+                    </>
+                  ) : null}
 
-              <Button
-                aria-label={
-                  user.authenticated && localCommunityJoined ? "Joined" : "Join"
-                }
-                color="white"
-                bg="black"
-                _dark={{ bg: "white", color: "black" }}
-                _hover={{
-                  bg: "brand.secondary",
-                }}
-                size="lg"
-                isLoading={localCommunityLoading}
-                onClick={() => {
-                  !user.authenticated
-                    ? (setAuthModal((prev) => ({
-                        ...prev,
-                        openAuthModal: true,
-                      })),
-                      toast({
-                        title: "You are not logged in!",
-                        description:
-                          "You are not allowed to join communities unless you log in",
-                        status: "error",
-                        duration: 2500,
-                        position: "bottom-left",
-                        isClosable: true,
-                      }))
-                    : handleJoinCommunity();
-                }}
-              >
-                {user.authenticated ? (
-                  localCommunityJoined ? (
-                    <Text fontSize="md">Joined</Text>
-                  ) : (
-                    <Text fontSize="md">Join</Text>
-                  )
-                ) : (
-                  <Text fontSize="md">Join</Text>
-                )}
-              </Button>
+                  <Button
+                    aria-label={
+                      user.authenticated && localCommunityJoined
+                        ? "Joined"
+                        : "Join"
+                    }
+                    color="white"
+                    bg="black"
+                    _dark={{ bg: "white", color: "black" }}
+                    _hover={{
+                      bg: "brand.secondary",
+                    }}
+                    size="lg"
+                    isLoading={localCommunityLoading}
+                    onClick={() => {
+                      !user.authenticated
+                        ? (setAuthModal((prev) => ({
+                            ...prev,
+                            openAuthModal: true,
+                          })),
+                          toast({
+                            title: "You are not logged in!",
+                            description:
+                              "You are not allowed to join communities unless you log in",
+                            status: "error",
+                            duration: 2500,
+                            position: "bottom-left",
+                            isClosable: true,
+                          }))
+                        : handleJoinCommunity();
+                    }}
+                  >
+                    {user.authenticated ? (
+                      localCommunityJoined ? (
+                        <Text fontSize="md">Joined</Text>
+                      ) : (
+                        <Text fontSize="md">Join</Text>
+                      )
+                    ) : (
+                      <Text fontSize="md">Join</Text>
+                    )}
+                  </Button>
+                </Flex>
+              </Flex>
             </Flex>
           </Flex>
         </Flex>
-      </Flex>
-    </Flex>
+      ) : communityExists === "no" ? (
+        <CommunityNotFound />
+      ) : communityExists === "unknown" ? (
+        <CommunityPageLoading />
+      ) : null}
+    </>
   );
 };
